refactor(dashboard): move the spend bar chart next to the usage page
