+++ web/apps/dashboard/app/(app)/[workspaceSlug]/settings/usage/spend-bar-chart.tsx
@@ -4,7 +4,7 @@ import { ChartContainer, ChartTooltip } from "@/components/ui/chart";
 import { formatPrice } from "@/lib/fmt";
 import { Skeleton } from "@unkey/ui";
 import { Bar, BarChart, CartesianGrid, XAxis, YAxis } from "recharts";
-import { ChartError } from "./components/chart-error";
+import { ChartError } from "@/components/charts/components/chart-error";
 
 export type SpendBarSeries = { key: string; label: string; color: string };
 export type SpendBarPoint = { time: number } & Record<string, number>;
